Implemented cw721 querying

+++ contracts/cw721-nfc/src/msg.rs
@@ -29,5 +29,5 @@ pub struct Cw721AddressResponse {
 // We define a custom struct for each query response
 #[derive(Serialize, Deserialize, Clone, Debug, PartialEq, JsonSchema)]
 pub struct GetCw721TokenNumResponse {
-    pub count: i32,
+    pub count: u64,
 }
